Remove the "RIP" from Parse
This list is about what the Stacks have _tried_ to do as much as
it's about what they're _going_ to do. Parse was Facebook's attempt
to put their own silo on top of the other stacks' platforms and
edge them out of their own devices. This is also true of Facebook's
Hello app.

Parse seems to have failed, but we can expect them to try again.